+++ src/Core.lua
@@ -43,7 +43,8 @@ function Yipper:OnEvent(event, ...)
                 ["NotificationSound"] = nil,
                 ["NotificationColor"] = Yipper.Constants.NotificationColor,
                 ["ShowHeader"] = true,
-                ["PingTrackedPlayer"] = false
+                ["PingTrackedPlayer"] = false,
+                ["EnableMinimapButton"] = true
             }
         -- If the YipperDB variable has been loaded for the character,
         -- assign it to the internal DB so the AddOn has its settings